+++ src-docs/src/views/datagrid/datagrid_styling_example.js
@@ -43,6 +43,7 @@ import {
   EuiDataGridStyle,
   EuiDataGridToolBarVisibilityOptions,
   EuiDataGridToolBarAdditionalControlsOptions,
+  EuiDataGridToolBarAdditionalControlsLeftOptions,
 } from '!!prop-loader!../../../../src/components/datagrid/data_grid_types';
 
 const gridSnippet = `<EuiDataGrid
@@ -98,20 +99,36 @@ const controlsSnippet = `<EuiDataGrid
   toolbarVisibility={{
     // Use of a fragment for multiple items will insure proper margins
     additionalControls: {
-      left: (
-        <Fragment>
-          <EuiButtonEmpty
-            size="xs"
-            onClick={() => {}}>
-            New button
-          </EuiButtonEmpty>
-          <EuiButtonEmpty
-            size="xs"
-            onClick={() => {}}>
-            Another button
-          </EuiButtonEmpty>
-        </Fragment>
-      ),
+      left: {
+        prepend: (
+          <Fragment>
+            <EuiButtonEmpty
+              size="xs"
+              onClick={() => {}}>
+              New button
+            </EuiButtonEmpty>
+            <EuiButtonEmpty
+              size="xs"
+              onClick={() => {}}>
+              Another button
+            </EuiButtonEmpty>
+          </Fragment>
+        ),
+        append: (
+          <Fragment>
+            <EuiButtonEmpty
+              size="xs"
+              onClick={() => {}}>
+              New button
+            </EuiButtonEmpty>
+            <EuiButtonEmpty
+              size="xs"
+              onClick={() => {}}>
+              Another button
+            </EuiButtonEmpty>
+          </Fragment>
+        ),
+      },
       right: (
         <Fragment>
           <EuiToolTip content="Right-side button">
@@ -261,22 +278,31 @@ export const DataGridStylingExample = {
         <>
           <p>
             Use the <EuiCode>toolbarVisibility.additionalControls</EuiCode> prop
-            to pass more buttons to the toolbar. It will respect the{' '}
-            <EuiCode language="js">toolbarVisibility={'{false}'}</EuiCode>{' '}
-            setting and hide when appropriate.
+            to pass more buttons to the toolbar.
           </p>
           <p>
             Passing a single node to <EuiCode>additionalControls</EuiCode> will
-            default to being appended to the left side of the toolbar. To
-            configure which side of the toolbar your controls display in, pass
-            an object with either the <EuiCode>left</EuiCode> or{' '}
-            <EuiCode>right</EuiCode> properties:
+            default to being placed in the <EuiCode>left.append</EuiCode>{' '}
+            position of the toolbar. To configure which side of the toolbar your
+            controls display in, pass an object with the <EuiCode>left</EuiCode>{' '}
+            or <EuiCode>right</EuiCode> properties:
           </p>
           <ul>
             <li>
               <EuiCode>additionalControls.left</EuiCode> appends the passed
-              custom control into the left side of the toolbar, after the column
-              & sort controls.
+              custom control into the left side of the toolbar.
+              <ul>
+                <li>
+                  <EuiCode>left.prepend</EuiCode> prepends the passed node into
+                  the left side of the toolbar, before the column & sort
+                  controls.
+                </li>
+                <li>
+                  <EuiCode>left.append</EuiCode> appends the passed node into
+                  the left side of the toolbar, after the column & sort
+                  controls.
+                </li>
+              </ul>
             </li>
             <li>
               <EuiCode>additionalControls.right</EuiCode> prepends the passed
@@ -299,6 +325,7 @@ export const DataGridStylingExample = {
         EuiDataGrid,
         EuiDataGridToolBarVisibilityOptions,
         EuiDataGridToolBarAdditionalControlsOptions,
+        EuiDataGridToolBarAdditionalControlsLeftOptions,
       },
       demo: <DataGridControls />,
     },